RavenDB-21797 Created separate handler for upgrade info and added better error handling
