+++ src/Raven.Server/Json/JsonDeserializationServer.cs
@@ -306,7 +306,7 @@ namespace Raven.Server.Json
 
         internal static readonly Func<BlittableJsonReaderObject, SubscriptionConnectionsDetails> SubscriptionConnectionsDetails = GenerateJsonDeserializationRoutine<SubscriptionConnectionsDetails>();
 
-        internal static readonly Func<BlittableJsonReaderObject, LicenseHandler.AboutViewResponse> AboutViewResponse = GenerateJsonDeserializationRoutine<LicenseHandler.AboutViewResponse>();
+        internal static readonly Func<BlittableJsonReaderObject, UpgradeInfoHandler.UpgradeInfoResponse> UpgradeInfoResponse = GenerateJsonDeserializationRoutine<UpgradeInfoHandler.UpgradeInfoResponse>();
 
         public sealed class Parameters
         {
